disable rendering while writing game attributes
if rendering is enabled during write operations the data will get
written to an unintended address resulting in graphical corruption.

add a missed case when writing game attributes, snap the screen
back to screen 0 after writing as a tradeoff between seeing screen 1
for a single frame vs a slight vertical misalignment for the same
frame.  not perfect, but less intensive than rewriting.

Signed-off-by: bunder2015 <omfgbunder@gmail.com>